Fixed input height and width sizes

+++ motor.cpp
@@ -3,7 +3,7 @@
 
 Motor::Motor() {}
 
-void Motor::setup(Ui::MainWindow *ui, uint8_t height, uint8_t width) {
+void Motor::setup(Ui::MainWindow *ui, uint16_t height, uint16_t width) {
   this->ui = ui;
   
   motor_fault = QPixmap(":/icons/motor.png");
